+++ src/arkode/arkode_impl.h
@@ -975,6 +975,7 @@ int arkSetInitStep(void *arkode_mem, realtype hin);
 int arkSetMinStep(void *arkode_mem, realtype hmin);
 int arkSetMaxStep(void *arkode_mem, realtype hmax);
 int arkSetStopTime(void *arkode_mem, realtype tstop);
+int arkClearStopTime(void *arkode_mem);
 int arkSetFixedStep(void *arkode_mem, realtype hfixed);
 int arkSetRootDirection(void *arkode_mem, int *rootdir);
 int arkSetNoInactiveRootWarn(void *arkode_mem);