+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # SUNDIALS Changelog
 
-## Changes to SUNDIALS in release 6.6.0
+## Changes to SUNDIALS in release 6.5.1
+
+Added the functions `ARKStepClearStopTime`, `ERKStepClearStopTime`,
+`MRIStepClearStopTime`, `CVodeClearStopTime`, and `IDAClearStopTime` to
+disable a previously set stop time.
 
 Fixed build errors when using SuperLU_DIST with ROCM enabled to target AMD GPUs.